backend/requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
sqlalchemy==1.4.41
pg8000==1.29.8
python-dotenv==0.21.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==1.10.7
httpx==0.23.3
python-multipart==0.0.5
